+++ tests/type-tests/test/orchestration.test-d.ts
@@ -275,6 +275,34 @@ expectError<any>(
 );
 
 /**
+ * Template role should be predefined string literals.
+ */
+expectError<any>(
+  new OrchestrationClient({
+    templating: {
+      template: [{ role: 'not-exist', content: 'Hello!' }]
+    },
+    llm: {
+      model_name: 'gpt-4o'
+    }
+  })
+);
+
+/**
+ * Tool message should have tool_call_id.
+ */
+expectError<any>(
+  new OrchestrationClient({
+    templating: {
+      template: [{ role: 'tool', content: 'Hello!' }]
+    },
+    llm: {
+      model_name: 'gpt-4o'
+    }
+  })
+);
+
+/**
  * Model parameters should accept known typed parameters and arbitrary parameters.
  */
 expectAssignable<LlmModelParams>({
@@ -303,12 +331,10 @@ expectType<Promise<OrchestrationResponse>>(
 );
 
 expect<ChatModel>('custom-model');
-expect<ChatModel>('gemini-1.0-pro');
 
 /**
  * Filtering Util for Azure content safety.
  */
-
 expectType<AzureContentSafetyFilterConfig>(
   buildAzureContentSafetyFilter({
     Hate: 'ALLOW_ALL',
@@ -328,7 +354,6 @@ expectError<AzureContentSafetyFilterConfig>(
 /**
  * Filtering Util for Llama guard.
  */
-
 expectType<LlamaGuard38BFilterConfig>(
   buildLlamaGuardFilter('code_interpreter_abuse', 'defamation')
 );