+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ai-sdk/core
 
+## 1.13.0
+
+### Minor Changes
+
+- a6ba3af: [Improvement] Add `gemini-2.0-flash`, `gemini-2.0-flash-lite`, `anthropic--claude-3.7-sonnet` and `deepseek-ai--deepseek-r1` to and remove `gemini-1.0-pro` from the available model list.
+
 ## 1.12.0
 
 ## 1.11.0