+++ packages/foundation-models/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ai-sdk/foundation-models
 
+## 1.13.0
+
+### Patch Changes
+
+- 8130838: [Fixed Issue] Remove incorrect error logging if finish reason is null in the streaming chunk.
+- Updated dependencies [a6ba3af]
+  - @sap-ai-sdk/core@1.13.0
+  - @sap-ai-sdk/ai-api@1.13.0
+
 ## 1.12.0
 
 ### Patch Changes